+++ api/helpers.py
@@ -1,7 +1,7 @@
-"""Helper modules for the Anker Power/Solix Cloud API."""
+"""Helper modules and classes for the Anker Power/Solix Cloud API."""
 
 from datetime import datetime, timedelta
-
+from cryptography.hazmat.primitives import hashes
 
 class RequestCounter:
     """Counter for datetime entries in last minute and last hour."""
@@ -37,3 +37,17 @@ class RequestCounter:
         """Get number of timestamps for last minute."""
         last_time = datetime.now() - timedelta(hours=1)
         return len([x for x in self.elements if x > last_time])
+
+
+def md5(text: str) -> str:
+    """Return MD5 hash in hex for given string."""
+    h = hashes.Hash(hashes.MD5())
+    h.update(text.encode("utf-8"))
+    return h.finalize().hex()
+
+
+def getTimezoneGMTString() -> str:
+    """Construct timezone GMT string with offset, e.g. GMT+01:00."""
+    tzo = datetime.now().astimezone().strftime("%z")
+    return f"GMT{tzo[:3]}:{tzo[3:5]}"
+
