+++ export_system.py
@@ -62,7 +62,7 @@ async def main() -> bool:
             )
             if resp != "" or not isinstance(randomize, bool):
                 randomize = resp.upper() in ["Y", "YES", "TRUE", 1]
-            nickname = myapi.nickname.replace(
+            nickname = myapi.apisession.nickname.replace(
                 "*", "x"
             )  # avoid filesystem problems with * in user nicknames
             folder = input(f"INPUT: Subfolder for export (default: {nickname}): ")
